requirements.txt
nltk==3.7
mammoth==1.5.0
python-docx==0.8.11
tqdm==4.64.1
word-forms==2.1.0
